erl2ocaml: map float() type into OCaml int to remove confusion
We already map float literals into int literals, so it is more straightforward if we map both types and values the same way.

+++ examples/dev/ir-spec/mod01.ml
@@ -251,7 +251,7 @@ let rec bitstring_id'1 : Ffi.bitstring'0 -> Ffi.bitstring'0 = function
 
 let rec byte_id'1 : Ffi.byte'0 -> Ffi.byte'0 = function v_A -> v_A
 
-let rec float_id'1 : float -> float = function v_X -> v_X
+let rec float_id'1 : int -> int = function v_X -> v_X
 
 let rec identifier_id'1 : Ffi.identifier'0 -> Ffi.identifier'0 = function
   | v_A -> v_A
